Extract MockUIPasteboard to a separate file
